Fix concurrent work
Replaces ThreadLocal with ConcurrentHashMap for application to be able
to close opened channels on the exit. Now it stores created channels in
values of Map, which can be iterated thru. In the shutdown hook I
iterate thru them and close each without handing exceptions (is it
correct to work this way?)